Name synth update handle

+++ src/main/audio/synth_internal.h
@@ -336,7 +336,7 @@ u32 synthSetHandleControllerValue(u32 handle, u8 controller, u8 value);
 u32 synthSetHandleControllerValue14Bit(u32 handle, u8 controller, u32 value);
 void synthCopyHandleFXState(u32 dstHandle, u32 srcHandle);
 u32 synthHandleKeyOff(u32 handle);
-void synthUpdateHandle(u32 value0, u32 value1, u32 handle, u8 mode);
+void synthUpdateHandle(u32 value0, u32 value1, u32 handle, u32 mode);
 u32 synthCancelCallbackVoices(u32 callbackId);
 
 #endif
